+++ test/DeckTest.java
@@ -16,14 +16,14 @@ public class DeckTest
 	{
 		Deck regular = new Deck();
 		assertEquals(52, regular.getSize());
-		Deck custom = new Deck(2);
+		Deck custom = new Deck((byte) 2);
 		assertEquals(104, custom.getSize());
 	}
 	
 	@Test(expected=IllegalArgumentException.class)
 	public void testForInvalidSize()
 	{
-		Deck bad = new Deck(-1);
+		Deck bad = new Deck((byte) -1);
 		assertEquals(-52, bad.getSize());
 	}
